requirements.txt
requests
python-dotenv
streamlit
matplotlib
streamlit-folium
folium